Evaluate the potentials at the collision center
The potentials of the incoming and outgoing particles used to be evaluated at
their positions respectively. But the outgoing particles are located at the
collisional center, which is different from the positions of the incoming
particles. This will lead to a difference on the potentials of the incoming
and the outgoing particles, which is an artifact. So now all the potentials
are evaluated at the collision center.

Some printing sentences are added for debugging. They shall be removed in the
final version

Signed-off-by: Feng Li <fengli@fias.uni-frankfurt.de>