Let the action finders know the potentials
I realize that I should pass the potential information to the actions
via the action finders, where the actions are initially created.
Before, the potentials were passed to the actions in the perform_action
function. Then the add_process function didn't know the potentials, so
those sub-threshold processes couldn't be rejected in advance.

Signed-off-by: Feng Li <fengli@fias.uni-frankfurt.de>