Rolled back changes to weight calculation for dileptons
With the output changes also the calculation of the dilepton
weights were changed that rendered them incompatible with the
analysis and also introduced a distinction that did not make
sense. Now the total weight represents the shining weight and
the partial weight is the same as for the normal decays.

refs #5643