Modify the test to be compatible with new config.yaml
It tourned that some of the tests relied on certain parameters being
accessible in the default config.yaml. Those were previously read
direcrly from the config. Now, the same parameters are initialised
direcrly in the tests.

Another remark: Jan and I found a parameter in the sanity and
initial_condition tests called "Sqrts_Reps" which does not appear
anywhere else in the code (neither as a switch nor a parameter) and is
not even used in the tests. We believe this is an artifact of some
previous implementation and removed it.