Disable strings for tests
The strings are for the moment disabled in the tests as they produce
memory leaks that are probably related to Pythia (not SMASH) and result
in some of the tests failing. See #5691. The plan for the future is to
find a way around this problem and in principle enable strings for the
test setups.