Remove photon cross sections from collision finding
Photon cross sections need not be added to the total cross section to
perform the collision finding. Such a treatment would contradict the
perturbative treatment that we apply for photons.
Additionally, I changed the value the cross section is set to whenever
the calculated cross section is <0. The encountered negative cross
sections are due to numerical errors. The exactly calculated values lie
between 0.01 and 0.7 in the sqrt_s range of interest. For the moment
I chose xsection = 0.1. But this is up for discussion.