Extend PDG codes to allow hexadecimal digits for J
This commit implements support for hex PDG codes, adds a few tests and
corrects the spin of the particles mentioned below.

N(2220), N(2250) and Λ(2350) have J=9/2, which cannot be represented in
a PDG code, because 2J+1=10 > 9. Extending the PDG codes to allow
hexadecimal digits (by allowing a-f as digit) makes them representable
with minimal changes to SMASH, because we use hex internally anyway.

However, this means these particles cannot be represented as a decimal
PDG code anymore, which has consequences for the binary and ROOT output.
For now, a PDG code of 0 is returned, when no decimal representation
exists. This is unfortunately not unique. The OSCAR output uses the
correct string representation, so now you might get letters in your PDG
codes, which will probably break unsuspecting scripts.

Refs #3872.