String cross section moved to cross_section class
The string cross section processes are now also calculated in
the cross section class. For this the decision of the soft
string subprocesses had to be moved to within the string
cross section function and is now stored in StringProcess,
where it can be accessed later when it is needed.

refs #5247