Increase pi+ p cross section
It was lower than before, mostly because we use more physical values for
the Δ(1950) branching ratios. Before, the N pi channel was much too
large (compared to the PDG branching ratio), however it described the
pi+ p cross section better. As a compromise, the N pi channel was
increased to the PDG maximum for all Delta resonances, describing the
cross section a bit better without goind into unphysical branching
ratios.

Refs #5773.