Decrease pi- p cross section
It was too high due to recent changes. By only modifying an N*
resonance, this will not decrease the pi+ p cross section, which is a
bit low. The Lambda producing channel was rescaled to make sure Lambda
production is not effected by this change. However, double pion
production may be effected.