Fixed mistake in nn matrix element calculation
The arguments were wrong before, since it was wrongly assumed
that always the matric elements for the incoming particle
types are used as arguments.
While changing this function, also the sqrt_s_ is now passed
as an argument so it now is static as it was before in the olc
scatteractionnucleonnucleon.cc file.

refs #5247