create a boolean fuction called "include_in_string"
to return whether the two species can interact via string. This
function is needed in both the cases: either turn on string with
a probability smoothly increasing from 0 to 1, or count the difference
between the parametrized total cross section and the total cross
section counted without string as the cross section of string
fragmentation.

Signed-off-by: Feng Li <fengli@fias.uni-frankfurt.de>