Implement two ways to switch on strings:
One is switching on the string fragmentations according to a
probability smoothly increasing with the colliding energy. The other
is always switching on the string fragmentations as long as the
parametrized total cross section is larger than the sum of the cross
sections of the non-string processes, with a cross section equal to
the difference between the previous two.

Signed-off-by: Feng Li <fengli@fias.uni-frankfurt.de>