Remove include assert.h statement in unittest.h
Master does not compile on Mac as long as assert.h is included in
unittest.h. Probably due to some wrong ordering or double inclusion of the header file. The error message in assert.h,
line 24 was thrown.
In fact, it is not necessary at all to include
assert.h in unittest.h. After removing it, SMASH compiles on Mac and all
tests pass on the FIAS machine.