Include assert.h again in unittest.h
This kind of reverts commit 1d0ea0d7. Here, the include statement in unittest.h was removed to make compilation on mac possible
Dima found that this introduced additional warnings during compilation, so I included this statement again - but with a different ordering of the include statements. Now the warnings disappear and SMASH also compiles on mac without errors.