Fix kaon charge exchange
Before the isospin ratios were calculated like for K N -> K Delta, which
was incorrect. Now they are not calculated, because the are one anyway.
This is verified by a newly added test.

To avoid issues like this in the future and to improve the confidence in
the code, the documentation of KplusNRatios has been improved and some
incorrect comments have been fixed.

Refs #6365.