Add hadrons
Some hadrons are included in the updated Pythia but not included
in the current SMASH yet. Most of them carry at least one bottom
quark. The PDG codes are taken from 2016 review of the particle
data group, and most of the masses are taken from ParticleData.php
under the Pythia folder.

Signed-off-by: Feng Li <fengli@fias.uni-frankfurt.de>