Implement Lorentz force
See Note #6189. The equations of motion are changed. As a consequence
of Lorentz boost, A Lorentz style force is acting on the baryons.

I also changed the way of calculating the derivaties on the lattices.
Before, the derivatives are estimated by the differences between the
values at the neighboring nodes. Now they are evaluated directly from
the Gaussian direvatives. This is the best way to evaluate the time
derivative. Otherwise one need to record the lattice values at last
time step and then take the differences between the values at the current
and last time step

Signed-off-by: Feng Li <fengli@fias.uni-frankfurt.de>