Include check for negative square of 4-vector.
Sometimes, the square of a 4-vector returns -0.0 instead of 0.0, which
results in a crash when taking the square root. Therefore, the absolute
value is taken to circumvent these crashes. To not miss future bugs with
negative values far smaller than 0.0, this only happens while the
deviation is smaller than really-small from zero.