Swap pion and rho mass if necessary in photon routine
t is (in the Turbide framework) defined to be the momentum exchange
between the rho and the photon. As the photon is the 4th particle, the
rho meson needs to be the second particle and thus m2 = m_rho. The order
of the incoming particles is however ambiguous, so that m2 could either
be the pion or the rho mass. Swapping of these masses needs therefore be
performed, if necessary.