Add parity to particles.txt
This cannot be derived from the PDG codes in general and has to be
specified separately. Knowing the parity is useful to check for parity
conservation in strong decays.

The values were taken from the PDG review. Parity for antiparticles is
the same for bosons and inverted for fermions.

Fixes and refs #2390.