Added config file for Travis CI
Travis CI is continious integration service, which
directly work with github. Everytime the master
gets updated all the tests are run automatically.
The same applies for all pull requests. Also, the
status of the last build is displayes within the
README on Github.

The travis.yml file is essentially the script
that runs everything. First it loads all
dependcies and installs them and then it
builds SMASH. At last the tests are run.

This intial setup only includes a simple setup
where only gcc is tested. This can be extended
in the future. We should probably also check
clang-format.

See travis.org for more details.

refs #1