Decision on allowed failure and notifications
In #1 we decided that mac has to pass compilation and test in
the future and we do not want to have an exception. Also
notification should go primarily to the committer. Since both,
notification of committer and mailing list, is not possible,
we will go with the travis default. The current notification
settings in travis.yml are only for testing purposes and will
be removed once this is done.