New helper script for clang-format
The idea is two-fold here: 1. the little helper script makes
it easier to run clang-format (use -p for that). No need to copy
and paste anymore. 2. with the -t option it can be used to test
if clang-format was run, which is important for continious
integration.

So far I only tested it on mac, will check also on linux.