Partially revert inheritance for deformed nuclei.
The routine for reading parameters for deformed nuclei from the config
was not working as expected. The constructor for a deformed nucleus called
the nucleus constructor and was inhereted. Here, only the parameters for
the Woods Saxon distribution were set, but not the ones specific for the
deformation. The ones setting the deformation were inherited from the
nucleus class but do actually initialize different setups. The latter
only sets the parameters for the spherical shape of the nuclei and their
orientation in space. I decided to remove the inheritance for these
functions and renamed them to point out that we are dealing with two
different kinds of initializations.

Refs #7150