Disable container overflow errors for mac
The AddressSanitizer has an know issue with false positive
container overflow errors, when different parts of an
application are built with and without sanitizer instrumentation.
See
https://github.com/google/sanitizers/wiki/AddressSanitizerContainerOverflow
for details.

For us the problem were yaml and pythia. We do not want (the user)
to addtionaly compile instrumented version of both. The issue
so far only hit on mac with AppleClang, therefore we disabling
it only there.

The allowed failure for Travis can with that also be removed.

This refs #6256.