Unify binaryoutput header
Before the binaryoutput header was split into collisions and
particle output. The binaryoutput base class was defined in the
binaryoutout collisions header and then this was imported into
the partciles header, which is I think very ugly and hard to
understand, when reading the code.

Now, only one header is used for binaryoutput, where all three
classes (base, collisions, partciles) are defined. The two .cc
are kept as is.

This finishes #6106