Unify binary output .cc files
While working with it, I noticed that is also cleaner to merge
the .cc files. For instance, the binaryoutput base functions and
the documentation where also (of course) just in the one .cc
files creating again a weird assymetry between them. 2 files
for the 3 classes were just a weird choice. Having one file
like for the other outputs is much cleaner.

This refs #6106