Fix the pi pi low energy elastic cross-sections
Turns out all meson meson-interactions in SMASH currently actually have
an AQM contribution. It was only disabled for the pion case because we
actually have data for these (at least for some of these). Since this
was already the case, this fix is the less invasive alternative to what
Vinzent suggested at the group meeting (checking if there are resonant
contributions). This basically modifies the exception so that pi+pi+ and
pi-pi- can now interact. If you think Vinzent's idea is more appropriate
I can have a look, but its more involved.

I also checked the behavior of the cross-section at very low p_lab
values, turns out it was kinda exploding up to ~27 mb. I introduced a
new function for the meson meson case which now checks the p_lab, and if
it's below a certain value, it will just use a flat 5 mb (well
technically 7.5 * 2/3, cause I thought that was clearer and didn't
obscure the AQM calculation)