Some changes to the user guide and doc
Added a footer to all documentation and user guide pages. We link now an
impressum and a data protection agreement. Since we host our pages on
GSI sites, we link to their impressum.

Worked on a small introduction to the user guide and documentation
pages. This needs certainly more work.