Changing only two body decay threshold
The previous change by Feng broke the dilepton run test.
We therefore settled on a smaller change by just touching the
threshold of the two body decay, which failed in #116. This
change fixes the issue, but does not break the dilepton run
test. We have to carefully look where else this replacment
makes sense (will be discussed in #76).