Improve smash -S close to the threshold
Before, we were neglecting that branching ratios in SMASH are
energy-dependent, leading to incorrect results close to the thresholds,
where some channels are not open. Now we consider the threshold for each
decay. To avoid numerical integration, we assume that all resonances are
at their pole mass. This is only an approximation, but it seems to work
well enough (the results of smash -S are now closer to actual SMASH
simulations).

Fixes #148.