Move 63-bit random seed generation to random.cc
This is very useful, if SMASH is used as external library. Then
instead of inventing the wheel, or copying existing smash code,
one can just call the smash function properly.

Also it makes code cleaner. Indeed, the technical implementation
of 63-bit random seed generation did not belong to smash.cc.
It makes much more sense to have it in the random:: namespace.

Signed-off-by: Dmytro Oliinychenko <doliinychenko@lbl.gov>