Fix unused values report and tests
Remove particles and decaymodes from Configuration,
where they do not need to be. The only reason to have them
there was to print them to the output config.yaml.

But now I simply print particles and decaymodes out explicitly,
because they are available as strings.

Also a configuration test was assuming that particles
and decaymodes are inside, so I had to fix it.

This refs #164

Signed-off-by: Dmytro Oliinychenko <doliinychenko@lbl.gov>