Increase error tolerance of histogram test
Sometimes the random tests fail, so the error tolerance was increased to
make that less likely to happen.

Our histogram error estimates are not very accurate, because we assume
a symmetric error distribution, but the Poisson distribution is
asymmetric. It is possible to use better estimates [1], but they are a
lot more complicated than the ones we are currently using, so relaxing
the error tolerance seems like the better option.

Fixes #133.

[1] http://articles.adsabs.harvard.edu/pdf/1986ApJ...303..336G