Improve QuantumNumber conservation check for high energies
In the case of high energy collisions (i.e., sqrts > 10 GeV),
the relative error in the four-momentum is also checked.
The four-momentum is considered to be conserved if the deviation
of each component is small enough compared to the invariant mass,
even though its absolute value is larger than small_number.