Fixed compilation when compiling as non-git repository
Our cmake build setup was relying on knowing the current git
branch, that is outputed for the `-v` command line option.

These changes remove the requirement of the source to be a git
repoisitory (the existence of a .git directory). Therefore, now
also the tarball one can download from github compiles without
problems.