Implement support for correlated nuclei
The CorrelatedNucleus class loads initial positions for nucleons from a
list into smash. These initial positions differ from the Woods-Saxon
ones due to the consideration of short range correlations between the
nuclei. It is now possible to use the new feature in the collider modus
setting: "Correlated: True", "File_Directory: <file directory>" and
"File_Name: <file name>" for the particle and the projectile.