Add additional condition to is_kinematically_possible
* It is not only sufficient to encounter sqrt(s) > 2*m_rho_minimum, in order to
create a rho meson in the final state, but we aso need to have remaining
energy in order to assign momentum to the final state particles.
Therefore, sqrt(s) needs to be at least slightly higher than the mass
threshold.
* The threshold for the mass is in fact not provided by 2*m_pion, but by
the lowest possible rho mass correspondig to a non-zero contribution of
the spectral function. In fact this value is slightly above the two-pion
mass. It can be accessed through min_mass_specral()
* Add test to ensure processes with sqrt(s) > 2*min_rho_mass, but
sqrt(s) < 2*min_rho_mass + really_small are not performed as there is no
energy remaining to assign momentum.

Fixes #140