Fix a bug related to grid + long timestep
The bug was showing very clearly, when SMASH is run as afterburner at LHC
energies. If one sets a large timestep, say, 1000 fm/c, then no collisions or
decays are found in the first 1000 fm/c.

This is because only particles, that are formed at the beginning of the
timestep were put on the grid for collision and decay search. It is obviously
wrong, because particles might get formed during the timestep.  So I had to
check cross-section scaling factor not in the beginning of the timestep, but in
the end of the timestep.

I am not convinced yet, that the bug in completely fixed, because for long
timesteps I see some increase of collision rate in the beginning of the next
timestep.  This is suspicious.

This refs #179

Signed-off-by: Dmytro Oliinychenko <doliinychenko@lbl.gov>