Change L of d->NN decay back to 1
This is necessary to reproduce the results of the
[arXiv:1809.03071], where d->NN decay had angular momentum 1.
Changing angular momentum to 0 was done, when parity was introduced.
This influenced the cross-sections.

Here I change the angular momentum back to 1. I have checked that the
cross-sections of the mentioned paper are reproduced. To fix parity violation,
it was enough to change d' parity from + to -. As d' is an artificial construct,
it's parity is not limited by physics and - is as good as +.

Signed-off-by: Dmytro Oliinychenko <doliinychenko@lbl.gov>