Fixed three-body-decay, see #32
The angular distribution of the three body decay was previously wrong. A
new algorithm takes care of this. Also the routine was renamed to
sample_3body_phasespace to be in line with the 2body routine.

Renamed one_to_three to sample_3body_phasespace to be in line with
sample_2body_phasespace. sample_3body_phasespace is now virtual, such
that it can be overwritten in decayactiondilepton.