Check deprecation of c function with cmake option (#195)
The deprecate_c_functions header was more or less unused up till
now, because it needed to be copy and pasted into all files
by hand. Now, it the deprecation warnings are easily accessible
by setting the cmake option DEPRECATE_C_FNS to ON. It is OFF
by default. The usage is specified in CONTRIBUTING.md.

* Do not assert test, if DEPRECATE_C_FNS. This is necessary, since otherwise code does not compiles, when deprecating c functions.
* Also, removed all present direct C functions calls