ryu/lund leading baryon (#201)
update in StringProcess class

In the non-diffractive soft string processes, the Lund fragmentation function
(with different parameters a and b) is used for the leading baryon fragmented
from a baryonic or anti-baryonic string.
When a separate fragmentation function for the leading baryon kicks in,
StringProcess::fragment_string keeps fragmenting hadrons until a baryon
(anti-baryon) is fragmented from the diquark (anti-diquark) end of a string,
or energy (mass) of the original string is used up.
If there is a remaining mesonic string, it is passed to PYTHIA fragmentation
routine.
It is possible to change StringZ::aLund and StringZ::bLund for the leading
baryons from the config file. (See Fragmentation.xml for more details about
StringZ::aLund and StringZ::bLund.)
Also StringFlav:popcornRate can be changed from the config file.
This popcorn rate determines the probability to have a meson from
the diquark end of a string. (See FlavorSelection.xml for more details about
StringFlav:popcornRate.)