Fix compiler warning about reaching end of non-void function
Although our case statements are exhaustive as proven by the compiler,
casting an integer to an invalid enum still might make the switch not
match on a case. Doing this would be very questionable, but it is not
clear whether the C++ standard allows this or not. We guard against this
case by throwing an exception, and thus fixing the warning.

Fixes #213.