Fix typo in deformed Woods-Saxon distribution
A set of parenthesis was missing in the implementation of the
Woods-Saxon distribution for deformed nuclei. As a consequence, only the
modified theta and phi-dependent radius was divided by the
diffusiveness, instead of the difference between radius and this
modified radius. See the SMASH paper, Eq. 21 for the exact formula.