Sanity check for pythia include
If two pythia versions are installed on a system, configuring SMASH with
a non-empty build folder can lead to a situation where the pythia
config-executable and the pythia library path do not match. This will
lead to a runtime crash.

With this fix we check inside Findpythia.cmake wether the two pathes
coincide.