Get rid of resonances in final state output
Currently, we assume for the final-state cross-section output that the
resonances have their pole mass. Due to this, we sometimes encounter
resonances that cannot decay, because the available energy is not high
enough to provide their pole mass.

This is a shortcoming of our simplifying assumptions. To get rid of the
resonances in the final state, this commit assigns zero cross section
for any reaction that is not possible with "mass = pole mass" and
filtering out cross sections that are zero everywhere.