Added Xi decays to N*
In order for this to work the N* pole masses also have to be
increased slightly within the pdg limits.

Similar idea found here: arXiv:1503.07305

Notes: One of the check involving clebsch gordon coefficents needs
to be commented out, since it is not compatible with the added
N* Xi decays.