Fix automatic initialization of deformed nuclei.
Apparently a field is removed from the YAML node once it is read in.
Taking the same argument twice does not work. The boolean parameter
whether or not the initialization of deformed nuclei happens
automatically or not is once read in in collidermodus.cc and again in
deformednucleus.cc. However in the latter, it always returns false since
the field seems to be empty. It thus needs to be fixed such that the
field in the YAML node is only read in once and the information passed
as an argument.