Pass grid cell volume to collision check function
The grid cell volume is now passed from the grid through
experiment to the scatteractionsfinder to check collisons. This
is necessary for the stochastic collision criterion.

Unfortunately, this change also caused changes to the other
action finders due to the inhertiance. Only the scatteractions-
finder needs the grid cell volume.