Add default value for hypersurface proper time and fix several bugs
* The default corresponds to the passing time when the nuclei have
entirely passed through each other
* Make SMASH crash if initial conditions should be extracted from
simulation in any other modus than collider
* Verify that all particles are removed at the same proper time (add
check)
* Make use of config.read() to read, but not take proper time argument
when initializing the hypersurfacecrossing action finder.

Bugfixes:
* Fix bug if proper time of particle is a NAN: This happens if t^2 < z^2, such that the value below the square root
becomes negative. Then, the comparison with a given proper time is not
well defined.
* Bugfix: Make sure, all particles are removed when crossing the
hypersurface
It was only assured that t0, which marks the beginning of the timestep, is
> 0. Particles could however have a positive t_end at the end of the
timestep corresponding to a tau > proper_time_of_hypersurface.
It is therefore necessary to assure that t_end > 0, regardless of the
value of t0