Avoid crashing, when HadgasEos solver is stuck
While solving the EoS of hadron resonance gas, one may realize, that
there is a reagion of (e, nb, ns), where there is no solution
for T, muB, muS. From time to time this region is touched by cells
of the grand-canonical thermalizer. In this case the hadgas solver
has no way to know that there is no solution and it crashes
after being unable to find any for a given number of iterations.

Crashing in this case is not justified. It typically happens in cases,
when the rest frame energy density is low enough, so the thermalizer
is not going to touch these cells anyway. Not crashing allows
to run large numbers of events with thermalizer without accidents.

Signed-off-by: Dmytro Oliinychenko <doliinychenko@lbl.gov>