Thermalizer: allow user to (un)choose microcanonical thermalization
In most of the use cases microcanonical thermalization is not really
needed. At the same time it slows simulation a lot. The microcanonical
option is now off by default, so the thermalizer can be run rather fast
with the biased Becattini-Ferroni algorithm.

This simple change makes the thermalizer much more usable in general.

Signed-off-by: Dmytro Oliinychenko <doliinychenko@lbl.gov>