Dump SMASH particle table in iSS format
This is very useful for consistent running of SMASH as afterburner
with CLVisc and MUSIC hydrodynamic codes. Both use the same
format of the particle table for sampling particles.

It is desirable that the sampled particles are consistent with
SMASH hadronic equation of state. This can now be achieved by
running

./smash -x > iSS_sampled_particle_list.txt

and using iSS_sampled_particle_list.txt as an input for the sampler.

Signed-off-by: Dmytro Oliinychenko <doliinychenko@lbl.gov>