Introduce parent class Clock
The functionality of the old Clock class is moved to the UniformClock
class, which is derived from Clock.
The clocks used in SMASH are now handled by unique pointers.
This is the prerequisite to introducing a second daughter class of clock
for custom output times.
Most unused functions of the clock are removed.