Enable frozen Fermi motion for hypersurface crossing actions
With frozen Fermi motion, the particles are propagated along the beam
direction without considering Fermi momenta until their first
interaction. The velocities and momenta stored in the particle data are
however already corrected for Fermi motion. Consequently, the action
finding needs to be performed with the beam velocity, not with the
corrected velocity. As auch, the beam momentum needs to be passed to the
find_actions_in_cells() method such that the properties are available.
It is unused in any other subclass.

Further Code clean-ups

* Remove unused parameters from override functions.
* Improve documentation
* Add documentation for beam_momentum parameter in find_actions_in_cell()