Fix action finding if frozen Fermi motion is enabled
The action finding with frozen Fermi motion enabled was previously performed with
the momentum of the particles (which already contains the Fermi motion
contributions). As the particles are however propagated with the beam
momentum along the beam axis until they interact (not with the momentum corrected by Fermi
motion), the action finding should also be performed with the momentum
of propagation, aka the beam momentum.

Fixes #329