Tune transition region for NN scatterings
The transition region was shifted from 4-5 GeV to 3.5-4.5 GeV to better
fit the experimentally measured excitation functions (<N>, <pT>, <mT>)
but at the same time to reproduce the exclusive cross sections. In here,
one is caught in a dilemma: Moving the transition region to
lower values significantly improves the excitation functions (dents,
kinks or similar are way smoother), but at the same time the cross sections for
exclusvive resonance production are cut off significantly, such that
those resonances can barely be produced anymore.
We have looked at the corresponding exclusive cross sections, the dilepton
production and the Pion, Kaon and Lambda excitation functions to find a
suitable compromise for this transition region. Letting the resonances
slowly die after sqrt(s) = 3.5 GeV still allows for all exclusive cross
section peaks to be matched (it just falls off a bit too sharp towards
higher energies). At the same time we managed to improve the excitation
functions by reducing the dents in 4pi-multiplicity and making the kink in
pT and mT spectra less pronounced. The remaining kinks are an artefact
of our transition treatment to move from a resonance description to a
string description. They cannot be reduced further unless cutting off a
significant contribution in the exclusive cross sections.

Fixes #130