Mohs/custom output times (#326)
Add an option to explicitly specify output times in the configuration file.

This involves making Clock an abstract class and add the derived classes UniformClock and CustomClock. The functionality of the old Clock is now found in UniformClock. The lab clock_, which is used for time steps of the simulation is now a std::unique_ptr<UniformClock>, while the outputclock is a std::unique_ptr<Clock>, so it can refer to both types of clocks, depending on the configuration.

A lot of unused functionality of the clock class is removed, but the possibility of adapting the time step is kept.
A bug is fixed, where the output time of the final output was not properly displayed on the terminal.
A debug output showing the wrong starting time of the output clock is removed.