Fix type confusion with global pointers
It turns out that the unused variable warning was probably not a false
positive, there was actually a type mismatch. I'm not sure whether that means
that the "potentials affect thresholds" option was completely broken!

Because their type was wrong, the definitions in `potential_globals.cc` were
effectively ignored, causing the lattice pointers to never be initialized,
likely resulting in undefined behavior.