Introduced a number of checks and warnings.
Introduced a check that the density lattice is used before calculating mean
field. Introduced a warning if potentials are used but lattice is not used
(so that the user knows mean fields are not going to be calculated).

Introduced a check that a theoretical prediction for the mean field is only
calculated in the Box Modus (otherwise it makes no sense).

Added description to calculate_mean_field_energy().

Used proper logging everywhere.

Moved the declarations of format_measurements() and
calculate_mean_field_energy() so that they are declared before they are
used.