New way of comparing current mean field to expectation, cosmetic changes.
1) Introduced a new way of judging the deviation of the mean field from
the expected value. Now the expected value is the value of the mean field
at t=0 instead of the theoretical prediction for the mean field in a
uniform box in rest frame. This also reduces the amount of computations
necessary. The new way is cleaner and more logical, and it has been tested
to trigger logging messages when the system deviates from equilibrium
(undergoes a phase transition).

2) Removed the "naive 2" way of calculating the mean field. The remaining
one are:
"naive 1": captures the way it should work for the Skyrme potentials, but
is only really correct in the rest frame.
"correct": is correct in any frame, but is not consistent with the way
SMASH currently calculates the EOMs.

3) Cleaned up comments, applied clang-format.