Remove assert statement in sample_resonance_mass()
The reason is that in DEBUG mode, neither the photon test passes, nor
the photon framework works. This is because of the perturbative
treatment of photons. Here, the sampling of the resonance masses in the
final state is performed before the kinematics are checked in the case
of photon processes. Therefore this function is sometimes called in
physically impossible cases (which are later on rejected). In those
cases, the assert statement would not be fulfilled and the execution of
SMASH is therefore aborted, although there is no actual problem.