Prelimiary implementation of HepMC output
* New HepMC output class and files
* Linking HepMC works, but paths have to specified by hand
in src/CMakeLists.txt
* Needs more work for master (see TODO comments) e.g. no
option for HepMC output (it is always produced)

Note, the output is most probable not in the format needed
for Rivet. Currently, only one vertex is created that
connects inital and final state. Also, only a minimum
amount of metadata is specified. This version serves
as a starting point to fine tune the output with
the experimentalists.

Refs #411