Add saveguard if cross section is negative.
If the sqrts values of the pion collision are below or above the values
provided in the tabularisation, the cross sections are yielded from
extrapolations at the edges. This may lead to negative cross sections
for very small values of sqrts. To prevent crashes those cross sections
are then manually set to the smallest value from the tabularisation.