New action class for three particle scatterings
Since the scatteraction class depends on two particles as incoming, a
new class to deal with three particles reactions is needed. It only
implements the necessary (pure virtual) functions. At the moment, it is
only used for known final states for testing purposes. Later, it can be
generalized to n-to-m reaction for the stochastic criterion.