Prevent FPEs if initial conditions output is enabled
* The modified terminal output prints the energy and the energy
difference normalized to the number of particles in each time step
* Running SMASH with the initial conditions enabled, this causes FPEs
because one would eventually divide by 0, once all particles have crossed
the hypersurface (ant were therefore removed)

The fixes are therefore:
* Add safeguard in case there are no more particles: kinetic and total
energy are set to 0.0
* The total energy difference is printed instead of the total energy
difference per particle. This is because the total energy difference per
particle cannot be computed if there are no remaining particles. A
corresponding warning is printed at the beginning of each new event.

* In addition, a runtest is added for an initial conditions run to catch
these problems earlier in the future.

Fixes #431