Use OutputOnlyFinal enum in root and binary output
Use the new enum OutputOnlyFinal that enables the option to only output
the final particles if the event was not empty in the root and binary
output code.

Also the output parameters adopt the enum and the config read in for the
enum is also added.

Note: This code version is not yet tested.