Fix cpplint on travis quickly
pip is no longer available out of the box on travis. pip3
is. So we use pip3 to install cpplint and add the PATH
respective install path to $PATH.

Might not be the ideal solution, since this might break again
soon. But it works for now and we need a quick solution for
the upcoming tag.