Add safeguard in case kinematics of collision do not allow production of a photon that is in accordance with the cross section cut-off
The cross section cut-off in the low-k direction is chosen to be k =
0.001. It is possible to have a pion-pion collision with very small
sqrt(s), where the largest possible value for the photon momentum is below this
cutoff. To make sure no photon is produced in such cases, the momentum
of the sampled photon is set to 0.0. In addition, the k-range that is
taken into consideration to determine the weighting factor used to be
negative, such that the square root could not be taken. This is however
equally wrong since this photon is assigned 0 momentum, it does not
exist and therefore needs to have a weight of 0 as well.