Scale stochastic criterion with cross section scaling factors
The probability for the stochastic collision criterion was previously 
calculated with the unscaled cross section, which is not correct.

Note that since the stochastic criterion was limited to the (elastic) 
box, no results were effected by this mistake, yet.

Also now static_cast testparticles in probability calculation for better 
style.