Consider formation times in photon production
The formation times of the particles and accordingly the cross section
scaling factor were not considered for the production of photons. The
weight was scaled with the underlying hadronic cross section (which was
scaled by the scaling factors), however the photon cross section was not
scaled accordingly. As a consequence, the weight for some photon
processes exploded as the underlying hadronic cross sections were tiny
whilst the photon cross sections remained unchanged (and unscaled). This
is now fixed such that also the photon cross sections are scaled accordingly.

Refs #427