Fix Inspire script
Due to the redesign of the Inspire website, the script to look up the references in SMASH for generating the BibLaTeX references was broken. For now, this can be fixed by explicitly requesting the old website. However, this is only a temporary measure, because the old website will eventually become inaccessible. Unfortunately, adapting the script to parse the new website is not feasible, because it uses a lot of JavaScript, making the generated BibLaTeX references hard to access.

Ideally, we should use the Inspire API instead, which is easier and more robust than parsing the HTML. However, the new API is not yet online, so this has to be deferred for now. I contacted the Inspire support and they told me that the old website will be accessible until the new API is available